Version from acer

+++ TSP/BFS.h
@@ -6,6 +6,6 @@
 class BFS
 {
 public:
-	static long long bfs(Vertex *vStart, std::unordered_set<Vertex*> &vertsLeft, std::vector<Vertex*> &path);
+	static double bfs(Vertex *vStart, std::unordered_set<Vertex*> &vertsLeft, std::vector<Vertex*> &path);
 };
